Move old LoginViewModel to splash directory
It's still used for the splash screen.
